[Front] 키보드 숏컷으로 다크모드 (ctrl + D)
